KAFKA-15774: refactor windowed stores to use StoreFactory
