+++ streams/src/main/java/org/apache/kafka/streams/kstream/internals/TimeWindowedKStreamImpl.java
@@ -33,18 +33,10 @@ import org.apache.kafka.streams.kstream.Window;
 import org.apache.kafka.streams.kstream.Windowed;
 import org.apache.kafka.streams.kstream.Windows;
 import org.apache.kafka.streams.kstream.internals.graph.GraphNode;
-import org.apache.kafka.streams.processor.internals.StoreBuilderWrapper;
-import org.apache.kafka.streams.processor.internals.StoreFactory;
-import org.apache.kafka.streams.state.StoreBuilder;
-import org.apache.kafka.streams.state.Stores;
-import org.apache.kafka.streams.state.TimestampedWindowStore;
-import org.apache.kafka.streams.state.WindowBytesStoreSupplier;
 import org.apache.kafka.streams.state.WindowStore;
 
-import java.time.Duration;
 import java.util.Objects;
 import java.util.Set;
-import org.apache.kafka.streams.state.internals.RocksDbIndexedTimeOrderedWindowBytesStoreSupplier;
 
 import static org.apache.kafka.streams.kstream.internals.KGroupedStreamImpl.AGGREGATE_NAME;
 import static org.apache.kafka.streams.kstream.internals.KGroupedStreamImpl.REDUCE_NAME;
@@ -113,7 +105,7 @@ public class TimeWindowedKStreamImpl<K, V, W extends Window> extends AbstractStr
 
         return aggregateBuilder.build(
             new NamedInternal(aggregateName),
-            materialize(materializedInternal),
+            new TimeWindowStoreMaterializer<>(materializedInternal, windows, emitStrategy),
             new KStreamWindowAggregate<>(
                 windows,
                 materializedInternal.storeName(),
@@ -165,7 +157,7 @@ public class TimeWindowedKStreamImpl<K, V, W extends Window> extends AbstractStr
 
         return aggregateBuilder.build(
             new NamedInternal(aggregateName),
-            materialize(materializedInternal),
+            new TimeWindowStoreMaterializer<>(materializedInternal, windows, emitStrategy),
             new KStreamWindowAggregate<>(
                 windows,
                 materializedInternal.storeName(),
@@ -216,7 +208,7 @@ public class TimeWindowedKStreamImpl<K, V, W extends Window> extends AbstractStr
 
         return aggregateBuilder.build(
             new NamedInternal(reduceName),
-            materialize(materializedInternal),
+            new TimeWindowStoreMaterializer<>(materializedInternal, windows, emitStrategy),
             new KStreamWindowAggregate<>(
                 windows,
                 materializedInternal.storeName(),
@@ -237,68 +229,6 @@ public class TimeWindowedKStreamImpl<K, V, W extends Window> extends AbstractStr
         }
         this.emitStrategy = emitStrategy;
         return this;
-    }
-
-    private <VR> StoreFactory materialize(final MaterializedInternal<K, VR, WindowStore<Bytes, byte[]>> materialized) {
-        WindowBytesStoreSupplier supplier = (WindowBytesStoreSupplier) materialized.storeSupplier();
-        if (supplier == null) {
-            final long retentionPeriod = materialized.retention() != null ?
-                materialized.retention().toMillis() : windows.size() + windows.gracePeriodMs();
-
-            if ((windows.size() + windows.gracePeriodMs()) > retentionPeriod) {
-                throw new IllegalArgumentException("The retention period of the window store "
-                        + name + " must be no smaller than its window size plus the grace period."
-                        + " Got size=[" + windows.size() + "],"
-                        + " grace=[" + windows.gracePeriodMs() + "],"
-                        + " retention=[" + retentionPeriod + "]");
-            }
-
-            switch (materialized.storeType()) {
-                case IN_MEMORY:
-                    supplier = Stores.inMemoryWindowStore(
-                        materialized.storeName(),
-                        Duration.ofMillis(retentionPeriod),
-                        Duration.ofMillis(windows.size()),
-                        false
-                    );
-                    break;
-                case ROCKS_DB:
-                    supplier = emitStrategy.type() == StrategyType.ON_WINDOW_CLOSE ?
-                        RocksDbIndexedTimeOrderedWindowBytesStoreSupplier.create(
-                            materialized.storeName(),
-                            Duration.ofMillis(retentionPeriod),
-                            Duration.ofMillis(windows.size()),
-                            false,
-                            false
-                        ) :
-                        Stores.persistentTimestampedWindowStore(
-                            materialized.storeName(),
-                            Duration.ofMillis(retentionPeriod),
-                            Duration.ofMillis(windows.size()),
-                            false
-                    );
-                    break;
-                default:
-                    throw new IllegalStateException("Unknown store type: " + materialized.storeType());
-            }
-        }
-
-        final StoreBuilder<TimestampedWindowStore<K, VR>> builder = Stores.timestampedWindowStoreBuilder(
-            supplier,
-            materialized.keySerde(),
-            materialized.valueSerde()
-        );
-
-        if (materialized.loggingEnabled()) {
-            builder.withLoggingEnabled(materialized.logConfig());
-        } else {
-            builder.withLoggingDisabled();
-        }
-
-        if (materialized.cachingEnabled()) {
-            builder.withCachingEnabled();
-        }
-        return new StoreBuilderWrapper(builder);
     }
 
     private Aggregator<K, V, V> aggregatorForReducer(final Reducer<V> reducer) {
